+++ manager/src/main/java/com/mucommander/ui/macosx/OSXIntegration.java
@@ -18,6 +18,9 @@
 
 package com.mucommander.ui.macosx;
 
+import com.mucommander.Launcher;
+import com.mucommander.commons.file.AbstractFile;
+import com.mucommander.commons.file.FileFactory;
 import com.mucommander.commons.runtime.OsFamilies;
 import com.mucommander.commons.runtime.OsVersions;
 import com.mucommander.conf.MuConfigurations;
@@ -26,8 +29,13 @@ import com.mucommander.conf.MuPreferences;
 import com.mucommander.ui.action.ActionManager;
 import com.mucommander.ui.dialog.about.AboutDialog;
 import com.mucommander.ui.dialog.shutdown.QuitDialog;
+import com.mucommander.ui.main.FolderPanel;
 import com.mucommander.ui.main.MainFrame;
 import com.mucommander.ui.main.WindowManager;
+import org.simplericity.macify.eawt.Application;
+import org.simplericity.macify.eawt.ApplicationEvent;
+import org.simplericity.macify.eawt.ApplicationListener;
+import org.simplericity.macify.eawt.DefaultApplication;
 
 /**
  * This class handles Mac OS X specifics when muCommander is started:
@@ -37,12 +45,11 @@ import com.mucommander.ui.main.WindowManager;
  *  <li>Registers handlers for the 'About', 'Preferences' and 'Quit' application menu items
  * </ul>
  *
- * <p>The com.apple.eawt API is used to handle 'About', 'Preferences' and 'Quit' events and report back to the OS.
+ * {@see Launcher} to find out it's usages
  *
- * @see EAWTHandler
- * @author Maxence Bernard
+ * @author Maxence Bernard, Eugene Morozov
  */
-public class OSXIntegration {
+public class OSXIntegration implements ApplicationListener {
 
     public OSXIntegration() {
         if(OsFamilies.MAC_OS_X.isCurrent()) {
@@ -63,17 +70,17 @@ public class OSXIntegration {
             System.setProperty("apple.laf.useScreenMenuBar", ""+MuConfigurations.getPreferences().getVariable(MuPreference.USE_SCREEN_MENU_BAR,
                                                                                                  MuPreferences.DEFAULT_USE_SCREEN_MENU_BAR));
 
-            // Catch 'About', 'Preferences' and 'Quit' events
-            new EAWTHandler();
+            Application app = new DefaultApplication();
+            app.setEnabledAboutMenu(true); // Enable the 'About' menu item
+            app.setEnabledPreferencesMenu(true); // Enable the 'Preferences' menu item
+            app.addApplicationListener(this); // Register this ApplicationListener
         }
     }
 
-    /**
-     * Shows the 'About' dialog.
-     */
-    public static void showAbout() {
+    public void handleAbout(ApplicationEvent event) {
+        event.setHandled(true);
         MainFrame mainFrame = WindowManager.getCurrentMainFrame();
-        
+
         // Do nothing (return) when in 'no events mode'
         if(mainFrame.getNoEventsMode())
             return;
@@ -81,10 +88,8 @@ public class OSXIntegration {
         new AboutDialog(mainFrame).showDialog();
     }
 
-    /**
-     * Shows the 'Preferences' dialog.
-     */
-    public static void showPreferences() {
+    public void handlePreferences(ApplicationEvent event) {
+        event.setHandled(true);
         MainFrame mainFrame = WindowManager.getCurrentMainFrame();
 
         // Do nothing (return) when in 'no events mode'
@@ -94,18 +99,45 @@ public class OSXIntegration {
         ActionManager.performAction(com.mucommander.ui.action.impl.ShowPreferencesAction.Descriptor.ACTION_ID, mainFrame);
     }
 
-    /**
-     * Quits the application after displaying a confirmation dialog if it hasn't been disabled
-     * in the preferences. Return <code>true</code> if the operation has been aborted by user.
-     */
-    public static boolean doQuit() {
+    public void handleQuit(ApplicationEvent event) {
+        boolean handled = true;
+
         // Ask the user for confirmation and abort if user refused to quit.
-        if(!QuitDialog.confirmQuit())
-            return false;
+        if(!QuitDialog.confirmQuit()) {
+            handled = false;
+        }
 
         // We got a green -> quit!
         WindowManager.quit();
-                
-        return true;
+
+        // Accept or reject the request to quit based on user's response
+        event.setHandled(handled);
+    }
+
+    public void handleOpenApplication(ApplicationEvent event) {
+        // No-op
+    }
+
+    public void handleReOpenApplication(ApplicationEvent event) {
+        // No-op
+    }
+
+    public void handleOpenFile(ApplicationEvent event) {
+        // Wait until the application has been launched. This step is required to properly handle the case where the
+        // application is launched with a file to open, for instance when drag-n-dropping a file to the Dock icon
+        // when muCommander is not started yet. In this case, this method is called while Launcher is still busy
+        // launching the application (no mainframe exists yet).
+        Launcher.waitUntilLaunched();
+
+        AbstractFile file = FileFactory.getFile(event.getFilename());
+        FolderPanel activePanel = WindowManager.getCurrentMainFrame().getActivePanel();
+        if (file.isBrowsable())
+            activePanel.tryChangeCurrentFolder(file);
+        else
+            activePanel.tryChangeCurrentFolder(file.getParent(), file, false);
+    }
+
+    public void handlePrintFile(ApplicationEvent event) {
+        // No-op
     }
 }
